Following changes from emb-cicle to the other geometry functions. Some bugfixes (moving variables out of headers).
