+++ emb-compress.h
@@ -45,64 +45,6 @@ extern EMB_PRIVATE int EMB_CALL husCompress(unsigned char* _266, unsigned long _
 #define bufferSize (512)
 
 /*****************************************
- * HUS Expand/Compress Variables
- ****************************************/
-short* _163;
-short* _164;
-unsigned char* _165;
-unsigned char* _166;
-unsigned short _167[17];
-short _168;
-short _169;
-short _170;
-short _171;
-short _172;
-short _173;
-short _174;
-short _175;
-short _176;
-short* _177;
-unsigned char* _178;
-unsigned char* _179;
-unsigned char* _180;
-unsigned char* _181;
-unsigned short _182;
-unsigned short _183;
-unsigned short _184;
-unsigned short _185;
-unsigned short _186;
-unsigned short* _187;
-unsigned short* _188;
-unsigned short* _189;
-unsigned short* _190;
-unsigned short* _191;
-unsigned short* _192;
-unsigned short* _193;
-unsigned short* _194;
-unsigned short* _240;
-unsigned short* _241;
-short _243;
-unsigned short _244;
-unsigned char _245;
-short _246;
-int _531;
-unsigned long _533;
-unsigned long _534;
-
-int mStatus;
-int currentIndex;
-long remainingBytes;
-int inputSize;
-int inputBufferSize;
-int inputLength;
-unsigned char* outputArray;
-unsigned char* inputArray;
-unsigned char* inputBuffer;
-int currentPosition;
-int inputPosition;
-int outputPosition;
-
-/*****************************************
  * HUS Expand Functions
  ****************************************/
 extern EMB_PRIVATE void EMB_CALL husExpand_258(int _259, unsigned char* _260, int _261, unsigned short* _262, unsigned short _263);
