--- geom-line.h
@@ -1,19 +1,0 @@
-/* Computational Geometry for Lines */
-#ifndef LIB_C_GEOMETRY_LINE_H
-#define LIB_C_GEOMETRY_LINE_H
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-void getLineIntersection(double  lineAx1,    double  lineAy1,
-                         double  lineAx2,    double  lineAy2,
-                         double  lineBx1,    double  lineBy1,
-                         double  lineBx2,    double  lineBy2,
-                         double *intersectX, double *intersectY);
-
-#ifdef __cplusplus
-}
-#endif /* __cplusplus */
-
-#endif /* LIB_C_GEOMETRY_LINE_H */
